move connection to AbstractTCP
